requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.0
geotext==0.4.0
git-filter-repo==2.47.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
